agent/pyproject.toml
[project]
name = "task-manager-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]
